moved scale_value to new file
